LPS-120269 Reorganize packages -test
