+++ modules/apps/portal-remote/portal-remote-cors-impl/src/test/java/com/liferay/portal/remote/cors/internal/url/pattern/mapper/factory/URLPatternMapperFactoryTest.java
@@ -12,7 +12,10 @@
  * details.
  */
 
-package com.liferay.portal.remote.cors.internal.url.pattern.mapper;
+package com.liferay.portal.remote.cors.internal.url.pattern.mapper.factory;
+
+import com.liferay.portal.remote.cors.internal.url.pattern.mapper.trie.DynamicSizeTrieURLPatternMapper;
+import com.liferay.portal.remote.cors.internal.url.pattern.mapper.trie.StaticSizeTrieURLPatternMapper;
 
 import java.util.HashMap;
 import java.util.Map;
